rework emu68 and display settings
