+++ docs/usage.md
@@ -8,9 +8,9 @@
 
 3. **Amiga Files tab - pick the Workbench version, icon set, and languages.** Add one or more directories that contain your Kickstart ROMs and Workbench ADFs. Click **Add...** to point at a folder; the list re-scans after each add/remove. Files are identified by hash, so filenames don't matter. Pick the **icon set** (GlowIcons on 3.2+, or Standard) next to the Workbench version, and tick any **language** disks you want installed. **Show details...** opens a tabbed view of what was detected: the boot ROM picked, the WHDLoad ROM inventory that gets staged to **DEVS:Kickstarts/**, and the per-ADF breakdown. For **AmigaOS 3.9**, add your OS 3.9 CD image (**.iso**, do not mount it) and a Kickstart 3.1 ROM instead of ADFs - the BoingBag updates download automatically.
 
-4. **Emu68 tab - pick the HDMI output and Workbench screen modes** and the Emu68 release to bundle (1.0.7 stable for pistorm32-lite / classic, or 1.1.0-alpha.1 for PiStorm16 support and the consolidated Pi4/CM4 driver stack - interrupt-driven ethernet, USB, NVMe). A VideoCore Workbench mode is written to **ScreenMode.prefs** during the build; native mode keeps the first-boot screen mode setup. Framethrower requires a VideoCore Workbench mode. The **Picasso96 RTG** box lets you point at your own commercial **Picasso96.lha**; left empty, the free Aminet version is used.
+4. **Emu68 tab - select the Emu68 release and boot behaviour.** Pick 1.0.7 stable for pistorm32-lite / classic or 1.1.0-alpha.1 for PiStorm16 support and the consolidated Pi4/CM4 driver stack. The basic view contains the CM4 antenna and storage timing. **Show Advanced** adds boot, memory, storage, compatibility, bus-test, and manual override settings. The right pane shows the generated **config.txt** and **cmdline.txt**, including values from the Display and Software tabs. With one-time diagnostics enabled, the affected file is split into first-boot and later-boot previews; the BAK file replaces the active file after the first boot.
 
-5. **Boot Files tab - configure Emu68 boot behaviour.** Pick the CM4 antenna, storage timing, and Framethrower settings. **Advanced Settings...** contains the low-level boot, memory, storage, and compatibility values. The selected Emu68 release determines whether compatibility, bus-test, and Framethrower options are written to **cmdline.txt** or as **config.txt** overlays. **Preview Files...** shows the rendered boot files without starting a build.
+5. **Display tab - configure HDMI and Workbench output.** Pick the Pi HDMI mode and Workbench screen mode, choose the Picasso96 source, and configure Framethrower / Unicam. A VideoCore Workbench mode is written to **ScreenMode.prefs** during the build; native mode keeps the first-boot screen mode setup. Framethrower requires a VideoCore mode. **Show Advanced** contains the HDMI hotplug setting.
 
 6. **Software tab - enable/disable optional packages.** Pick MUI 3.8 or 5.0, plus any applications and commodities you want. Mandatory system packages are always included and not shown here. The first build downloads everything (with caching, subsequent builds are faster).
