+++ src/main/python/emu68hatcher/data/packages/clicktofront.yaml
@@ -1,7 +1,7 @@
 name: clicktofront
 friendly_name: ClickToFront
 group: Commodities
-description: 'click a window with a qualifier to bring it to front; moved to WBStartup to run at boot'
+description: 'click to bring windows to front)'
 versions:
 - '3.1'
 - '3.2'
