+++ src/main/python/emu68hatcher/data/packages/hid2ami.yaml
@@ -1,7 +1,7 @@
 name: hid2ami
 friendly_name: HID2AMI Mouse Wheel Driver
 group: Drivers
-description: 'mouse wheel support for the HID2AMI usb mouse/gamepad adapter; needs firmware 3.0.0+'
+description: 'mouse wheel support (firmware 3.0.0+)'
 versions:
 - '3.1'
 - '3.2'
