+++ src/main/python/emu68hatcher/data/packages/poseidon.yaml
@@ -1,7 +1,7 @@
 name: poseidon
 friendly_name: Poseidon USB Stack
 group: Drivers
-description: 'USB support for the Pi4/CM4 ports via xhci.device - Poseidon stack with Trident prefs, xhci units preconfigured'
+description: 'USB support for Pi4/CM4 ports'
 versions:
 - '3.1'
 - '3.2'
