+++ src/main/python/emu68hatcher/data/packages/winzoom.yaml
@@ -1,7 +1,7 @@
 name: winzoom
 friendly_name: WinZoom
 group: Commodities
-description: 'arrange windows via hotkeys; runs from WBStartup'
+description: 'arrange windows with hotkeys'
 versions:
 - '3.1'
 - '3.2'
