+++ src/main/python/emu68hatcher/gui/dialogs/__init__.py
@@ -1,7 +1,6 @@
 """GUI dialogs"""
 
-from emu68hatcher.gui.dialogs.boot_files_preview_dialog import BootFilesPreviewDialog
 from emu68hatcher.gui.dialogs.build_progress_dialog import BuildProgressDialog
 from emu68hatcher.gui.dialogs.detected_files_dialog import DetectedFilesDialog
 
-__all__ = ["BootFilesPreviewDialog", "BuildProgressDialog", "DetectedFilesDialog"]
+__all__ = ["BuildProgressDialog", "DetectedFilesDialog"]
