+++ src/main/python/emu68hatcher/gui/tabs/__init__.py
@@ -1,6 +1,6 @@
 """GUI tab widgets for the main window"""
 
-from emu68hatcher.gui.tabs.boot_files import BootFilesTab
+from emu68hatcher.gui.tabs.display import DisplayTab
 from emu68hatcher.gui.tabs.emu68 import Emu68Tab
 from emu68hatcher.gui.tabs.kickstart import KickstartTab
 from emu68hatcher.gui.tabs.network import NetworkTab
@@ -12,8 +12,8 @@ from emu68hatcher.gui.tabs.start import StartTab
 __all__ = [
     "StartTab",
     "KickstartTab",
-    "BootFilesTab",
     "Emu68Tab",
+    "DisplayTab",
     "PackagesTab",
     "NetworkTab",
     "PartitionsTab",
